# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kubernetes资源管理API，支持动态创建开发环境和SA权限管理",
        "title": "K8S Resource API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/k8s",
    "paths": {
        "/environments": {
            "post": {
                "description": "基于ServiceAccount的kubeconfig创建完整的K8s环境，包括PVC、Deployment和Service。使用传入的kubeconfig中的身份信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "summary": "创建环境",
                "parameters": [
                    {
                        "description": "创建环境请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateEnvironmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            }
        },
        "/environments/list": {
            "post": {
                "description": "列出用户有权限访问的所有环境，支持分页和简单的命名空间过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "parameters": [
                    {
                        "description": "列出环境请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ListEnvironmentsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功，data中包含环境列表",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            }
        },
        "/environments/{name}": {
            "post": {
                "description": "获取指定名称的环境的状态信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "summary": "获取环境信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "环境名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "获取环境信息请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.GetEnvironmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定名称的环境，包括PVC、Deployment和Service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "summary": "删除环境",
                "parameters": [
                    {
                        "type": "string",
                        "description": "环境名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "删除环境请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.DeleteEnvironmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            }
        },
        "/service-accounts/list": {
            "post": {
                "description": "列出用户有权限查看的ServiceAccount，支持分页和命名空间过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service-accounts"
                ],
                "parameters": [
                    {
                        "description": "列出ServiceAccount请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ListServiceAccountsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功，data中包含ServiceAccount列表",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            }
        },
        "/service-accounts/{name}": {
            "post": {
                "description": "创建指定的ServiceAccount，分配权限，并返回对应的kubeconfig",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service-accounts"
                ],
                "summary": "创建ServiceAccount并返回kubeconfig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ServiceAccount名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "创建ServiceAccount请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.KubeconfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定的ServiceAccount及其相关权限配置（Role、RoleBinding、ResourceQuota）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service-accounts"
                ],
                "summary": "删除ServiceAccount",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ServiceAccount名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "删除ServiceAccount请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.DeleteServiceAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "main.CreateEnvironmentRequest": {
            "type": "object",
            "required": [
                "image",
                "kubeconfig",
                "name",
                "resources",
                "storage"
            ],
            "properties": {
                "image": {
                    "type": "string",
                    "example": "registry.km.top/kmai/ubuntu:22.04-ide"
                },
                "kmcode_config": {
                    "type": "object",
                    "additionalProperties": true
                },
                "kubeconfig": {
                    "type": "string",
                    "example": "apiVersion: v1\nkind: Config\n..."
                },
                "name": {
                    "type": "string",
                    "example": "demo"
                },
                "namespace": {
                    "type": "string",
                    "example": "dev-space"
                },
                "nodeports": {
                    "type": "object",
                    "properties": {
                        "ssh": {
                            "type": "integer",
                            "example": 0
                        },
                        "terminal": {
                            "type": "integer",
                            "example": 0
                        },
                        "vscode": {
                            "type": "integer",
                            "example": 0
                        }
                    }
                },
                "resources": {
                    "type": "object",
                    "required": [
                        "cpu",
                        "cpu_limit",
                        "memory",
                        "memory_limit"
                    ],
                    "properties": {
                        "cpu": {
                            "type": "string",
                            "example": "1000m"
                        },
                        "cpu_limit": {
                            "type": "string",
                            "example": "2000m"
                        },
                        "memory": {
                            "type": "string",
                            "example": "2Gi"
                        },
                        "memory_limit": {
                            "type": "string",
                            "example": "4Gi"
                        }
                    }
                },
                "storage": {
                    "type": "object",
                    "required": [
                        "vscode",
                        "workspace"
                    ],
                    "properties": {
                        "vscode": {
                            "type": "string",
                            "example": "5Gi"
                        },
                        "workspace": {
                            "type": "string",
                            "example": "10Gi"
                        }
                    }
                }
            }
        },
        "main.DeleteEnvironmentRequest": {
            "type": "object",
            "required": [
                "kubeconfig",
                "namespace"
            ],
            "properties": {
                "kubeconfig": {
                    "type": "string",
                    "example": "apiVersion: v1\nkind: Config\n..."
                },
                "namespace": {
                    "type": "string",
                    "example": "dev-space"
                }
            }
        },
        "main.DeleteServiceAccountRequest": {
            "type": "object",
            "required": [
                "namespace"
            ],
            "properties": {
                "kubeconfig": {
                    "type": "string",
                    "example": "apiVersion: v1\nkind: Config\n..."
                },
                "namespace": {
                    "type": "string",
                    "example": "dev-space"
                }
            }
        },
        "main.GetEnvironmentRequest": {
            "type": "object",
            "required": [
                "kubeconfig",
                "namespace"
            ],
            "properties": {
                "kubeconfig": {
                    "type": "string",
                    "example": "apiVersion: v1\nkind: Config\n..."
                },
                "namespace": {
                    "type": "string",
                    "example": "dev-space"
                }
            }
        },
        "main.KubeconfigRequest": {
            "type": "object",
            "properties": {
                "create_if_not_exists": {
                    "type": "boolean",
                    "example": true
                },
                "kubeconfig": {
                    "type": "string",
                    "example": "apiVersion: v1\nkind: Config\n..."
                },
                "namespace": {
                    "type": "string",
                    "example": "default"
                },
                "resource_limits": {
                    "type": "object",
                    "properties": {
                        "cpu": {
                            "type": "string",
                            "example": "8"
                        },
                        "memory": {
                            "type": "string",
                            "example": "16Gi"
                        },
                        "pod_count": {
                            "type": "string",
                            "example": "2"
                        },
                        "storage": {
                            "type": "string",
                            "example": "20Gi"
                        }
                    }
                },
                "token_expiration_hours": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "main.ListEnvironmentsRequest": {
            "type": "object",
            "required": [
                "kubeconfig"
            ],
            "properties": {
                "kubeconfig": {
                    "type": "string",
                    "example": "apiVersion: v1\nkind: Config\n..."
                },
                "namespace": {
                    "description": "可选，不指定则用kubeconfig中的",
                    "type": "string",
                    "example": "dev-space"
                },
                "page": {
                    "description": "页码，从1开始",
                    "type": "integer",
                    "default": 1,
                    "example": 1
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer",
                    "default": 20,
                    "example": 20
                }
            }
        },
        "main.ListServiceAccountsRequest": {
            "type": "object",
            "properties": {
                "kubeconfig": {
                    "description": "可选，不提供则用管理员kubeconfig",
                    "type": "string",
                    "example": "apiVersion: v1\nkind: Config\n..."
                },
                "namespace": {
                    "description": "可选，不指定则列所有命名空间",
                    "type": "string",
                    "example": "dev-space"
                },
                "page": {
                    "description": "页码，从1开始",
                    "type": "integer",
                    "default": 1,
                    "example": 1
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer",
                    "default": 20,
                    "example": 20
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}